Broke ExternalCompactionIT into 3 different IT classes
Broke ExternalCompactionIT into 3 different IT classes, also
moved the compaction related ITs to a compaction package
