+++ test/src/main/java/org/apache/accumulo/test/compaction/CompactionRateLimitingIT.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.accumulo.test;
+package org.apache.accumulo.test.compaction;
 
 import static org.junit.Assert.assertTrue;
